Downloads working with CountdownLatch + minor fixes

+++ src/Messages/FileBlockRequestMessage.java
@@ -17,7 +17,7 @@ public final class FileBlockRequestMessage implements Serializable {
         return blockOffset;
     }
 
-    public int getBlockLength(){
+    public long getBlockLength(){
         return blockLength;
     }
